+++ GeekBrains_Cource_Two/Controllers/FriendsViewController/FriendsViewControllerDelegate.swift
@@ -1,0 +1,9 @@
+//
+//  FriendsViewControllerDelegate.swift
+//  GeekBrains_Cource_Two
+//
+//  Created by Samborsky on 30.06.2022.
+//
+
+import UIKit
+
