+++ GeekBrains_Cource_Two/Controllers/GalleryViewController/GalleryViewController+DataSource.swift
@@ -1,0 +1,22 @@
+//
+//  GalleryViewController+DataSource.swift
+//  GeekBrains_Cource_Two
+//
+//  Created by Samborsky on 30.06.2022.
+//
+
+import UIKit
+extension GalleryViewController: UICollectionViewDataSource {
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        return photos.count
+    }
+
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: reuseIdentifierCustom, for: indexPath) as? GalleryCollectionCell else { return UICollectionViewCell()}
+
+        cell.configure(image: photos[indexPath.item])
+
+        return cell
+    }
+
+}
