+++ GeekBrains_Cource_Two/Controllers/GroupsViewController/GroupsViewController+FillData.swift
@@ -1,0 +1,12 @@
+//
+//  GroupsViewController+FillData.swift
+//  GeekBrains_Cource_Two
+//
+//  Created by Samborsky on 30.06.2022.
+//
+
+import UIKit
+
+
+///заполняем массив групп
+
