+++ GeekBrains_Cource_Two/Controllers/GroupsViewController/GroupsViewController.swift
@@ -6,14 +6,13 @@
 //
 
 import UIKit
-
 class GroupsViewController: UIViewController {
 
     @IBOutlet weak var groupsTableView: UITableView!
 
     let reuseIdentifierCustom = "reuseIdentifierCustom"
-    var groupsArray: [Groups] = []
     let FromGroupsToMygroups = "FromGroupsToMygroups"
+    var groupsArray: [Groups] = []
 
     func fillGroupArray() {
         let group1 = Groups(title: "ТВ передачи", avatar: UIImage(named: "group2"), description: "расписание телепередач на каждый день")
@@ -48,45 +47,4 @@ class GroupsViewController: UIViewController {
 
     }
 
-
 }
-//MARK: - расширения для таблицы UITableView вкладка Группы
-
-extension GroupsViewController: UITableViewDelegate, UITableViewDataSource {
-
-///метод позволяющий менять размер изображения в ячейке(высоту ячейки)
-    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        return 200
-    }
-
-    ///метод количества секций в таблице
-    func numberOfSections(in tableView: UITableView) -> Int {
-        return 1
-    }
-
-    ///метод количества строк в таблице
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return groupsArray.count
-    }
-
-    ///метод позволяющий заполнять ячейки таблицы
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        //приводим нашу ячейку к типу xib файла(UITableViewCell)
-        guard let cell = tableView.dequeueReusableCell(withIdentifier: reuseIdentifierCustom, for: indexPath) as?
-                CustomTableViewCell else { return UITableViewCell()}
-//        cell.configure(friend: friendsArray[indexPath.row]) // эта строка позволяет добавить все компоненты массива сразу (name, age, avatar) или можно как ниже, каждый добавить отдельно
-        cell.nameLabel.text = groupsArray[indexPath.row].title
-        cell.avatarImageView.image = groupsArray[indexPath.row].avatar
-        cell.ageLabel.text = groupsArray[indexPath.row].description
-        cell.viewForShadow.layer.shadowColor = UIColor.white.cgColor
-        return cell
-    }
-
-    //метод обрабатывающий нажатие на ячейку. можно вывести в консоль куда именно нажал человек
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        performSegue(withIdentifier: FromGroupsToMygroups, sender: groupsArray[indexPath.row])
-
-            print("нажата строка \(indexPath.row) в секции \(indexPath.section). Вкладка Группы")
-        }
-}
-
