+++ GeekBrains_Cource_Two/cells/GalleryCollectionCell.xib
@@ -24,7 +24,7 @@
                                 <rect key="frame" x="8" y="8" width="226" height="194"/>
                             </imageView>
                             <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="T28-u2-HS1" customClass="LikeControll" customModule="GeekBrains_Cource_Two" customModuleProvider="target">
-                                <rect key="frame" x="207" y="211" width="25" height="25"/>
+                                <rect key="frame" x="207" y="204" width="25" height="25"/>
                                 <subviews>
                                     <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="heart" catalog="system" translatesAutoresizingMaskIntoConstraints="NO" id="xBx-bF-ahl">
                                         <rect key="frame" x="0.0" y="1.5" width="25" height="22.5"/>
@@ -44,7 +44,7 @@
                                 </connections>
                             </view>
                             <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="UBN-Be-42D">
-                                <rect key="frame" x="197" y="236" width="0.0" height="0.0"/>
+                                <rect key="frame" x="207" y="206" width="0.0" height="0.0"/>
                                 <fontDescription key="fontDescription" type="system" pointSize="17"/>
                                 <nil key="textColor"/>
                                 <nil key="highlightedColor"/>
@@ -52,13 +52,13 @@
                         </subviews>
                         <color key="backgroundColor" systemColor="systemBackgroundColor"/>
                         <constraints>
-                            <constraint firstAttribute="bottom" secondItem="UBN-Be-42D" secondAttribute="bottom" constant="6" id="Ang-qg-CZj"/>
-                            <constraint firstAttribute="bottom" secondItem="T28-u2-HS1" secondAttribute="bottom" constant="6" id="E50-CG-fCD"/>
+                            <constraint firstItem="UBN-Be-42D" firstAttribute="top" secondItem="bet-qU-7Ra" secondAttribute="bottom" constant="4" id="8ha-P8-Hwh"/>
+                            <constraint firstItem="T28-u2-HS1" firstAttribute="leading" secondItem="UBN-Be-42D" secondAttribute="trailing" id="Wz5-Fn-tPw"/>
                             <constraint firstAttribute="bottom" secondItem="bet-qU-7Ra" secondAttribute="bottom" constant="40" id="keC-Yp-GFA"/>
                             <constraint firstAttribute="trailing" secondItem="T28-u2-HS1" secondAttribute="trailing" constant="10" id="msq-Bn-2JH"/>
                             <constraint firstItem="bet-qU-7Ra" firstAttribute="leading" secondItem="NpE-wh-FX5" secondAttribute="leading" constant="8" id="nF4-y6-mPQ"/>
+                            <constraint firstItem="T28-u2-HS1" firstAttribute="top" secondItem="bet-qU-7Ra" secondAttribute="bottom" constant="2" id="qs8-KZ-cUz"/>
                             <constraint firstItem="bet-qU-7Ra" firstAttribute="top" secondItem="NpE-wh-FX5" secondAttribute="top" constant="8" id="tAL-Fp-qsh"/>
-                            <constraint firstItem="T28-u2-HS1" firstAttribute="leading" secondItem="UBN-Be-42D" secondAttribute="trailing" constant="10" id="z8M-DG-8FI"/>
                             <constraint firstAttribute="trailing" secondItem="bet-qU-7Ra" secondAttribute="trailing" constant="8" id="zV9-4q-jyW"/>
                         </constraints>
                     </view>
